backend/clean_requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
asyncpg==0.30.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
crashtest==0.4.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
distlib==0.3.9
distro==1.9.0
dotenv==0.9.9
dulwich==0.22.8
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
fonttools==4.58.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.6
jiter==0.10.0
kaleido==0.2.1
keyring==25.6.0
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.7.0
msgpack==1.1.1
narwhals==1.40.0
numpy==2.2.5
ollama==0.4.8
openai==1.93.0
packaging==25.0
pandas==2.2.3
pbs-installer==2025.7.2
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.8
plotly==6.1.1
poetry==2.1.3
poetry-core==2.1.3
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==6.2.0
reportlab==4.4.1
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
tomlkit==0.13.3
tqdm==4.67.1
trove-classifiers==2025.5.9.12
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.35.0
vercel-ai==0.1.0
virtualenv==20.31.2
watchfiles==1.1.0
websockets==15.0.1
zstandard==0.23.0
